Make time parsing work in Safari
There are two issues:

1. Unlike Chrome and Firefox, Safari does not parse date-time strings like
   where date and time is separated by a space as in '2018-07-28 19:30'.
   However, as promised by ECMAScript 5 spec, it does parse date-time
   strings where the separator is 'T' like '2018-07-28T19:30'.
2. ECMAScript 5 spec says that a timestring above should be parsed as an
   UTC time. ECMAScript 6 says that it should be parsed as a local time.
   *Of course* Safari does the former and the other browsers do the
   latter.

This patch attempts to deal with both issues and parse 'YYYY-MM-DD
HH:MM' as a local time in all the browsers.

I couldn't find a compability table for the second issue, which I take
to mean that nobody uses Date.parse in anger. That's probably a good
idea.